refactoring onboarding in progress 1
